// src/app/interfaces/iloginconfig.ts
export interface Iloginconfig {
    ConfigID: any;
    BannerImage: any;
    BackgroundColor: any;
    TextColor: any;
    Alignment: any;
    ButtonColor: any;
    ButtonTextColor: any;
    Logo: any;
    LoginFormBackgroundcolor: any;
    CreatedBy: any;
    TenantID: any;


}
export interface IThemeconfig {
    ConfigID: any;
    Theme: any;
    HeaderLogo: any;
    NavBar: any;
    PrimaryBgColor: any;
    SecondaryBgColor: any;
    PrimaryTextColor: any;
    SecondaryTextColor: any;
    ActiveColor: any;
    FooterText: any;
    FavIcon: any;
    CreatedBy: any;
    TenantID: any;
}
export interface ITableconfig {
    ConID: any;
    Pagination: any;
    Scrollable: any;
    Sorting: any;
    Filter: any;
    ExportButton: any;
    MultiSelect: any;
    Resizable: any;
    Reorder: any;
    CreatedBy: any;
    TenantID: any;
}

export interface Iconfig {
    TenantId: any;
    LanguageId: any;
}

export interface ILoginuser {

    EmailID: any,
    Password: any

}